update
set up and tested register controller & route. Had issues testing if the controller worked and req.body being parsed for data to be saved in mongoDB. Turns out that raw was not set to json and it could not parse. alot of time wasted, but a good lesson to learn.