Update on users.js
added way to get all users and a way for users to delete their own information. Still need to add more functionality with the delete controller (as in we can delete the userId but the auth token associated with it still exist). Check comments within the user.js in controllers for more information