Small refactor and using pop up feature for user account handling
Small refactoring to handle the use of TextField component in CustomInput (it seems that it does not have placeholder property as it uses label prop for its placeholder)

Added pop up feature in index of landing page, I should make it a component for modularity. I am wondering how this would affect the toggle state if there are multiple pop up components.